requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
drfpasswordless==1.5.3
importlib-metadata==1.6.0
kombu==4.6.8
pytz==2019.3
redis==3.5.2
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
